pimd: Moving pim_cmd_lookup_vrf function to common file(pimd.c)
